words-backend/requirements.txt
django==4.1.3
pika==1.3.1
pika-stubs==0.1.3
psycopg2-binary==2.9.5
pyyaml==6.0